requirements.txt
# Minimum Django and REST framework version
Django>=2.2.1
djangorestframework>=3.9.4

# Test requirements
factory-boy==2.11.1
pytest-factoryboy==2.0.2
pytest==4.4.0
pytest-cov==2.6.0
pytest-django==3.4.3
pytest-sugar==0.9.1
pytest-xdist==1.29.0
ipdb==0.12
pylint==2.3.1
coreapi==2.3.3
autopep8==1.4.4


# wheel for PyPI installs
wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.11.1
